requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2==2.9.10
PyMySQL==1.1.1
pyspark==3.3.2
PyYAML==6.0.2
Werkzeug==3.1.3
